requirements.txt

python-telegram-bot==20.7
apscheduler==3.10.4
pytz==2024.1
sqlalchemy==2.0.23